pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "sdlg"
version = "0.0.1"
description = "A pure Python written library built on top of PySDL3 to create programs and games."
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE.txt" }
keywords = ["Python", "SDL3", "game development", "PySDL3", "pygame"]
authors = [{ name = "Emanuel Hesselroth", email = "emanuelodenhesselroth@gmail.com" }]
maintainers = [{ name = "Emanuel Hesselroth", email = "emanuelodenhesselroth@gmail.com" }]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3 :: Only"
]
dependencies = ['pysdl3', 'psutil']

[project.optional-dependencies]
dev = ["pytest"]
test = ["pytest"]

[project.urls]
"Homepage" = "https://github.com/EmanuelNorsk/sdlg"
"Issues" = "https://github.com/EmanuelNorsk/sdlg/issues"
"Source" = "https://github.com/EmanuelNorsk/sdlg/issues"

[tool.setuptools]
# If there are data files included in your packages that need to be
# installed, specify them here.
# package-data = { "sdlg" = ["*.dat"] }
# The package itself is automatically detected when using pyproject.toml.